+++ assets/javascripts/kitten/components/layer/modal-next/index.js
@@ -535,7 +535,7 @@ const Actions = ({
 
 const ModalButton = props => (
   <Button
-    size="big"
+    size="large"
     fit="fluid"
     {...props}
     className={classNames('k-ModalNext__buttons', props.className)}
@@ -705,7 +705,7 @@ const InnerModal = ({
                   <CloseButton
                     modifier="hydrogen"
                     onClick={close}
-                    size="regular"
+                    size="medium"
                     closeButtonLabel={closeButtonLabel}
                     aria-labelledby={
                       !!headerMessage ? 'ModalHeaderMessage' : null