+++ assets/javascripts/kitten/components/navigation/pagination/index.js
@@ -155,7 +155,7 @@ export const Pagination = forwardRef(
             key={`link-${number}`}
             className="k-Pagination__Link"
             tabIndex="0"
-            size="tiny"
+            size="small"
             fit="icon"
             {...buttonProps}
           >
@@ -200,7 +200,7 @@ export const Pagination = forwardRef(
             tabIndex={0}
             onClick={isDisabled ? null : pageClickHandler(number)}
             disabled={isDisabled}
-            size="tiny"
+            size="small"
             fit="icon"
           >
             <ArrowIcon