+++ assets/javascripts/kitten/karl/pages/news-list/index.js
@@ -109,7 +109,7 @@ const NewsList = () => {
       </Head>
       <Container>
         <NewsContainer>
-          <Button size="big" fit="fluid" modifier="helium">
+          <Button size="large" fit="fluid" modifier="helium">
             <InstrumentTagIcon width="14" />
             <span>Créer une nouvelle actu</span>
           </Button>